refactor:関心毎に分ける
